Fix failing emerge caused by ppc-openbsd
Since c605e11ea has been commited, compilers can't be emerged, no-one
knows why the line was added, but it is known to break everything.

```
localhost ~ # emerge -avq '=dmd-2.071.2'

!!! All ebuilds that could satisfy "=dmd-2.071.2" have been masked.
!!! One of the following masked packages is required to complete your request:
- dev-lang/dmd-2.071.2::dlang (masked by: invalid: DEPEND: USE flag 'ppc-openbsd' referenced in conditional '!ppc-openbsd?' is not in IUSE, invalid: RDEPEND: USE flag 'ppc-openbsd' referenced in conditional '!ppc-openbsd?' is not in IUSE)

For more information, see the MASKED PACKAGES section in the emerge
man page or refer to the Gentoo Handbook.
```

If you ever decide to add it back, explain in commits why! And test that
the packages actually build!