dmd.eclass: explicitly specify arguments to dmd in tests
Refactor the test code a little for better readability and prevent
linking by default to libphosbos2.a, which happens to work only because
the build wrongfully generates both shared and static libraries
regardless of USE=static-libs.

Pass the same options to dmd as the ones in the installed dmd.conf to
get a closer setup to what the user will have installed on their system.

Signed-off-by: Horodniceanu Andrei <a.horodniceanu@proton.me>