dev-lang/ldc2: remove github PRs from SRC_URI and put them in FILESDIR
pkgcheck has warned me that fetching them from github may lead to them
changing, which is exactly what happened leading to manifest failures.

I did it because I thought that, since the PR is closed and I'm the
author, the possibility of it changing is slim and the alternative was
to get a size violation warning from pkgcheck.

I have now learned why I shouldn't ignore pkgcheck errors and, if
commits need to be taken from upstream, I should keep their size to a
minimum and only backport commits that fix stuff, not commits which add
features.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>