dmd-r1.eclass: clarify dependency on stdlib of host compiler
The previous comment, saying that dmd doesn't link the standard lib of
the host compiler, is plain wrong, it always does. The only
particularity is that dmd doesn't seem to need phobos but it does need
druntime. Regardless, a RDEPEND on the host compiler is needed.

Another comment has been added trying to describe the reason why
DLANG_DEPS should be present in DEPEND. My current understanding is that
since the compiler needs to link libraries from SYSROOT (druntime and
phobos) to produce the CHOST dmd executable then those libraries should
be listed in DEPEND. I'm not 100% sure though.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>