dev-lang/ldc2: add 1.39.0
The LICENSE has been corrected as I've mistaken the license of the
valgrind headers as being BSD when, in fact, it is BZIP2. The other
change to LICENSE, RSA being replaced by `|| ( CC0-1.0 Apache-2.0 )' is
because some internal hashing implementation has changed.

The arm64 keyword has been dropped temporarily as I find myself without
the hardware to test it.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>