dev-util/gdmd: Drop untested keywords from older slots
gdmd used to inherit the KEYWORDS from the gcc:${SLOT} it
mirrored. There is little value in having this little script be
keyworded for arches that no-one is using gdc for and it only leads to
the KEYWORDS becoming out of date as old gcc versions are masked in
::gentoo.

Only keep a reasonable subset of the KEYWORDS, much like it is done in
newer versions of the package.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>