dev-util/gdmd: Replace SLOTs with USE flags
Replace the multiple dev-util/gdmd:1{2..5} with the single
dev-util/gdmd:0[dlang_targets_gdc-1{2..5}].

This helps with:
- updating the innate package, since there's only 1 ebuild now
- using it for older/newer gcc's through DLANG_COMPAT_OVERRIDE
- installing common files, e.g. the man page

The only downside is handling the conflicts since older D packages would
RDEPEND on gdmd:<SLOT> which gives annoying blocker messages when trying
to update. The issue would not manifest instantaneously, but when one D
package is updated it would then try to pull in gdmd:0 which would
conflict with the gdmd:<SLOT> that is still a dependency on other,
older, D packages. This can be improved by adding "virtual" packages in
the form:
- dev-util/gdmd:15 RDEPENDs on dev-util/gdmd:0[dlang_targets_gdc-15]

After enough time has passed and the D packages have been
rebuilt/updated, the alias packages can simply be dropped.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>