eclass: Drop ugly KEYWORDS abuse
The code paths are not useful for users, they're only meant for
developers to help with pkgcheck usage. The newly added profiles are a
better tool, being more expressive, not requiring code and not being
visible to users.

On top of this, because the eclasses were inherited before the KEYWORDS
were defined, the code wasn't doing anything.

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>